requirements.txt
Cython==0.29.17
matplotlib==3.2.2
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
PyYAML==5.3
scipy==1.4.1
tensorboard==2.1
torchvision==0.8.1
tqdm==4.41.0
seaborn==0.11.1
pandas==1.0.3
torch==1.7.0